Use Object.create to cheaply clone options, allowing mutation
By using Object.create, the original options become the prototype
of the newly created object, and so any mutation of the options
object should just shadow the properties of the prototype.

Essentially, this is like cloning by reference, rather than by
value, and has almost no performance penalties in this context.