Allows intercepting and skipping the AtRule node
While postcss-import crawls imports, it is often handy to be able to
mutate the rule before processing. Unfortunately I was unable to
accomplish this by using a plugin before postcss-import. This is due to
rules not being revisited after concatenated with postcss-import.

I also found it odd that there was no way to signal a skip for a given
rule. By signaling a `skip = true` on the node, postcss-imports safely
skips over, but emits a warning. I don't like mutating the AtRule
object with a custom skip property, but this works for now.